requirements.txt
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
tiktoken>=0.8.0
